move install_package to packages repository
