+++ src/server/api.rs
@@ -8,11 +8,11 @@ use serde::Serialize;
 use tokio::sync::Notify;
 
 use crate::common::semver::Semver;
+use crate::frontend::check_release_against_installed_package;
 use crate::frontend::pkg::repository::PackagesRepository;
 use crate::frontend::releases::{
   Release, Version, fetch_remote_frontend_package_release, get_remote_release,
 };
-use crate::frontend::{check_release_against_local_one, install_package};
 use crate::server::common::{ServiceError, empty_body, full_body};
 
 #[derive(Serialize)]
@@ -31,7 +31,7 @@ pub async fn check_latest_frontend_release()
 -> Result<Response<BoxBody<Bytes, ServiceError>>, ServiceError> {
   let response = match get_remote_release(Version::Latest).await {
     Ok(latest_release) => {
-      let (local_version, _) = check_release_against_local_one(&latest_release).await;
+      let (local_version, _) = check_release_against_installed_package(&latest_release).await;
       let response_body = CheckLatestResponseBody {
         should_update: local_version.is_none_or(|local| local < latest_release.version),
         latest_release,
@@ -88,7 +88,7 @@ pub async fn update_frontend_package(
   };
 
   const FORCE_OUTDATED: bool = true; // TODO: this should be provided from frontend. atm always force outdated pkg
-  match install_package(path, FORCE_OUTDATED, pkgs_repo).await {
+  match pkgs_repo.install_package(path, FORCE_OUTDATED).await {
     Ok(()) => {
       let response = Response::new(empty_body());
       Ok(response)
